pyproject.toml
[project]
name = "moltbook-analysis"
version = "0.1.0"
description = "Memetic and language-ontology analysis for Moltbook"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]

dependencies = [
  "httpx>=0.27",
  "tenacity>=8.2",
  "pydantic>=2.6",
  "pandas>=2.1",
  "pyarrow>=15",
  "tqdm>=4.66",
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "scipy>=1.11",
  "networkx>=3.2",
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "langdetect>=1.0.9",
  "python-dotenv>=1.0.1",
  "beautifulsoup4>=4.12",
  "lxml>=5.1"
]

[project.optional-dependencies]
dynamic = [
  "playwright>=1.41"
]

[project.scripts]
mbk = "moltbook_analysis.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
